Move usages to Usage/usage.ts
