+++ frontend/src/Part/GearCard.svelte
@@ -7,10 +7,10 @@
     Row,
     Col,
   } from "@sveltestrap/sveltestrap";
-  import { types, usages, fmtSeconds, fmtDate, fmtNumber } from "../lib/store";
+  import { types, fmtSeconds, fmtDate, fmtNumber } from "../lib/store";
   import { Part } from "../Part/part";
   import { link } from "svelte-spa-router";
-  import { Usage } from "../Usage/usage";
+  import { usages, Usage } from "../Usage/usage";
 
   export let part: Part;
   export let display = false;
